Bootloader/stage1: When parsing the memory map from BIOS, check that the map follows the assumptions made in this bootloader.
The BIOS function to get the memory map (INT=15h AX=E820h) does not
guarantee that the returned map is sorted nor does it guarantee that
ranges of the map are not overlapping.
In practice though, it seems that the map is always sorted and never
contains overlapping entries. To keep things simple, the bootloader
assumes this, for now. Hence, when parsing the map, check that those
assumptions hold and panic if that is not the case.
If in the future a machine/VM triggers this panic, the parsing code
would need to be changed to both sort the map and merge overlapping
ranges.