Bootloader/stage1: Improve LOG macro to put strings in .data section.
Prior to this commit the LOG macro would just insert the string passed
as argument in the middle of the code being generated and "jump" over it
before calling the printf32/64 function. This led to a lot of garbage in
the middle of good code where LOG was invoked in turn leading to garbage
output from disassemblers.