Bootloader/stage1: Add first steps to load the kernel image from disk.
Introduce the concept of "Metadata Sector" which is a special sector on
disk containing information on the kernel image such as its location on
disk as well as its size. This special sector is expected to immediately
follow the last stage1 sector on disk.
For now the loader is only capable of loading the Metadata Sector.