Bootloader/stage1: Reduce .text size by half by trimming some fat of the logging macros.
The logging macros used to insert code saving all regs on the stack
prior to the actual print code. This would lead to substantial code
bloat, particuarly in 64-bit mode, as 16 push and 16 pop instructions
would be inserted at each invocation of a logging macro.

The register saving has been implemented in the printf functions
instead, avoid a lot of code duplication. The resulting .text size has
be halved.