Bootloader/stage1: Reduce stage1's image size by avoid string duplication due to PANIC macro.
The PANIC macro used to invoke the _LOG macro to print the
banners/separators (e.g "=== .. PANIC .. ===" and "=== ... ==="). This
led to a lot of wasted space as those string would get duplicated in the
resulting image.

Changed the PANIC macro to use a helper function printing the
banners/separator so that the strings are no longer duplicated.