Kernel: Add definition of __cxa_atexit.
GCC expects this function to exists. It is responsible to register all
destructors that should be called when a shared library needs to be
unloaded. AFAIK there is no way to tell GCC not to use it (I did not
bother to search very much), defining it here seems easy enough ...