Bootloader/stage1: Introduce bootstruct and pass it to the kernel.
The bootstruct is a data structure created by the stage1 bootloader that
is passed to the kernel.
For now, this structure contains information about:
  - The e820 memory map.
  - The free-list of physical page frames.

Upon jumping to the kernel's entry point, the bootloader writes the
address of the bootstruct into the %RDI register.

Added necessary code to pass the pointer to the bootstruct to
kernelMain.