Kernel: Add Res<T> class.
A Res<T> contains either a value of type T or an error of type Error
(enum defined in error.hpp). Such a type is useful for functions that
need to return a value but also be able to indicate an error condition
if something goes wrong.
Added tests for this new type.