Kernel: Add Err type.
Err is a type representing an error or the lack thereof.
An Err optionally contains an error. It is used for functions that
typically return void, but still need to communicate whether or not an
error occured. For functions that need to both return a non-void value
_and_ communicate errors use Res<T> instead.
In case of an error an Err contains an Error value, otherwise it
contains no value.

Added tests for this new type as well.