EmbeddedFreeList: Fix two issues with allocation.
First, allocations cannot be smaller than sizeof(Node). This is because
in case such an allocation is freed, there is not enough space to create
a new Node. Therefore allocation smaller than sizeof(Node) should
actually allocate sizeof(Node) bytes. This is transparent to the caller
which is still expected to pass the original size (< sizeof(Node)) when
calling free().

Second, when allocating into a particular node, the resulting size of
the node should not become < sizeof(Node) as otherwise a new node cannot
be created. Therefore some allocations are not possible, for instance
trying to allocate 2 bytes in a node of size sizeof(Node)+1 is not
possible. Make sure the allocator refuses such allocations.

Added regression tests for both of the issues above.