Freelist: When allocating, allocate from the base of a freelist node not from the end.
Allocating from the end of a node means that two allocations A1 and A2
ends up with A2 < A1. This is a rather strange behaviour for an
allocator and can lead to more internal fragmentation when allocating
heap memory (heap allocation is still in progress at the time of this
commit).