Interrupts: Add support for custom interrupt handlers.
Using Interrupts::registerHandler(vector, func) one can register a
function `func` that will be called everytime an interrupt with vector
`vector` is raised. Conversely Interrupts::deregisterHandler(vector)
removes the currently registered handler for the vector.

This feature will prove itself useful for testing. For instance it can
be used to verify that a page-fault occurs under certain conditions and
"catch" that page-fault directly in the test.