SubRange: Add new type SubRange.
Inspired by Pascal, a SubRange is type capable of holding only values
contained in a certain sub-range of u64. Any value outside of the range
[Min; Max] is rejected by the constructor by a panic.
SubRange is NOT a collection, it is merely a wrapper around the
underlying type which performs range checks.
SubRange should not be used as-is but instead should be used as a base
type for a new type, for example:
        `class Ring : public SubRange<Ring, 0, 3> {};`
The Impl template parameter is used to make sure that the comparison
operators and copy constructor only take parameters of type Impl and not
SubRange<Impl,Min,Max>, this means that even if A and B are deriving
from SubType with same Min and Max, they cannot be compared or assigned
(assuming they set their Impl to A and B respectively).
SubType does not provide arithmetic operators such as +, -, *, ...
because it is expected that most types deriving from SubRange don't have
well-defined arithmetic operators (what does it mean to divide two
values of type Ring??). Nevertheless, if a particular type requires
those operators, it can perform those operations on the raw value using
raw().