Segmentation: Fix issue with far jumps when setting the CS register which led to a crash when running the kernel using KVM on AMD processors.
When attempting to run the kernel using qemu+KVM, the kernel would crash
on the far jump in _setCs but only when running on an AMD processor;
running on an Intel processor would run fine.
It turns out, AMD processors only support far jump to 32bit offsets in
64-bit mode and that offset is ZERO extended to 64-bit.  Since we are in
higher half, this cannot work (sign-extended would have worked). Our
only option is therefore to use a RETF, e.g. far return.  Although not
mentioned in the docs, it seems that RETF is capable of returning to a
far pointer of the form m16:64 (e.g. with 64-bit offset). My guess is
this is because a far call can be done from any canonical 64-bit offset
and as such the RETF must support jumping back to a 64-bit offset.
Intel CPUs do not have this issue since they support far jump to m16:64.
But keep it simple and use the RETF in both cases.
For both Intel and AMD it seems that a RETF pops 16 bytes from the
stack, e.g. 8-bytes for the offset + 8-bytes for the segment selector.
This is not very clear in the docs, so just to be safe the segment
selector is zero-extended before being pushed on the stack.