ACPI: First implementation of ACPI v1.0 table parsing.
This first implementation searches for the RSDP then parses the RSDT and
lists the SDTs it finds in the RSDT. For now it only logs of the SDTs it
finds, e.g. no information is saved. This first implementation contains
some warts that will be cleaned up in subsequent commits.

For now only ACPI revision v1.0 is supported as this is what is provided
by Qemu. AFAIK most non-ancient real hardware is using ACPI v2.0.
Support for v2.0 will be added later and is not expected to be a
subsequent change.