Paging: Better estimate the size of physical memory when creating the Direct Map.
The previous estimate would read the e820 memory map and compute the
size of physical memory as the highest offset of available memory. This
is incorrect as we might miss some reserved memory towards the end of
physical memory. This was actually the case for the ACPI tables under
Qemu: those tables are located at the very top of physical memory in RAM
marked as reserved in the last entry of the map. The previous estimate
would therefore skip this region of RAM when creating the Direct Map,
which meant the ACPI tables could not be read using the Direct Map.

This commit fixes the estimation to instead take the highest offset
listed in the e820 memory map even if it refers to reserved memory, but
there's a catch: Qemu on AMD hosts with IOMMU enabled!

It turns out that Qemu creates an entry in the e820 memory map when
running on AMD hosts with IOMMU enabled (see the pc_memory_init()
function in i386/pc.c). This entry is there to mark IOMMU memory as
reserved. However, such memory starts around offset 1TiB.
Therefore we must skip this entry when estimating the size of physical
memory otherwise we might try to create a Direct Map for 1TiB, most of
which is useless since there is no way the guest has that much memory.