ACPI: Add proper parsing for the MADT table.
For now the parser simply dumps the content of the MADT and its entries
into the console. Actually storing that information will be done in a
subsequent commit.

Also got rid of the hack mapping the RSDT to virtual memory. This is not
needed since @e9738e6 fixed the Direct Map being too short and not
containing the full physical memory.