I/O APIC: Fix two BUGs.
The first issue was RedirectionTableEntryReservedBits being wrong, it
marked the bits of the `destination` field of a redirection table entry
as reserved. This meant that the destination would always be set to 0 no
matter the RedirectionTableEntry.

The second issue was reserved bits clobbering: when writing an entry
into the redirection table, the impl would set all reserved registers to
0 instead of leaving them untouched.