PIT: Add support for the Programmable Interval Timer (PIT).
Added functions to configure the PIT:
        - Map the PIT IRQ to a particular vector
        - Configure the PIT frequency
        - Disable (e.g. mask) the PIT.

PIT code is kept pretty simple because the only planned use-case for it
is to compute the frequency of the LAPIC timer (to be done in a
subsequent commit).
The only PIT mode supported/used is "Mode 3 - Square Wave Generator"
which generates periodic interrupts at a configurable frequency.

Added testing as well.