LAPIC timer: Add LapicTimer::delay() function.
This new function takes a Duration as argument and delays the cpu by the
time indicated in the Duration. The delay is implemented as a busy-loop
counting the number of elapsed ticks on the LAPIC timer in order to
detect when the duration has elapsed.

This commit also adds Timer::Duration, a new type used to represent a
time duration with max precision of microseconds.