SMP/Test: Fix broken SMP test.
The wakeApplicationProcessorTest() was broken due to the fact that the
started AP would get the value of its IP register by making a call,
without having a stack setup. This lead to memory corruption.

Rewrote the startup code part of the test. It turns out we don't need
the IP as the IP is always 0x0 and the CS comes from the vector field of
the SIPI.