Misc: Re-organize includes to avoid cyclic dependencies.
Many headers, especially headers that define types, were un-necessarily
directly or indirectly including panic.hpp or log.hpp. This leads to a
lot of pain down the line as panic.hpp and log.hpp ultimately include
logger.hpp which needs to include the .hpp of the type it supports
printing.

Try to avoid this as much as possible, either by replacing PANIC by
ASSERT in some type definition, or, in the case of selftest.hpp, move
the macros using PANIC and LOG to their own header file that is included
by tests as needed.