Concurrency: Add Atomic<T> type to represent atomic values.
This new type allows to manipulate a value in an atomic fashion. For now
this first version only accepts unsigned integers as its template
parameter and all Atomic<T> are implemented as an atomic u64 to keep
thing simple.

Atomic<T> provides an interface to read, write, increment,
compare&exchange the underlying value.