Atomic: Add basic test for Atomic<T> operators and fix bug in Atomic<T>::operator-=(u64).
This test merely checks that the operators of Atomic<T> are correctly
implemented. It does NOT check that Atomic<T> is indeed atomic and
thread-safe, that will come in a later test...

This new test actually proved itself useful as it already uncovered a
bug in Atomic<T>::operator-=(u64).