SMP: Fix race-condition in AP wake-up code.
The finalizeApplicationProcessorStartup() had a rare race-condition due
to the way APs indicate when they are online.
Prior to this commit, AP would perform the following in order to jump to
their target 64-bit code:
        1. Notify the cpu doing the wakeup that this AP is online by
           setting apStartupFlag = 1.
        2. Switch to the newly allocated stack.
        3. Jump to the target code.
This, however, had a race condition: between steps 1 and 2, the AP still
uses its old stack. If another AP is started during that time, we end
with two APs using the same stack.

Hence use a trampoline which sets the apStartupFlag and jumps to the
target code _after_ switching to the new stack.

Added an overload for Stack::switchToStack() in which it is possible to
pass an argument to the target function.