Util: Add Ptr<T> class implementing smart pointers.
A Ptr<T> implements a smart pointer to a object of type T allocated on
the heap. A Ptr<T> keeps track of the reference count of the object it
points to. Copying a Ptr<T> creates a _new_ reference to that object,
increasing the reference count. Destroying a Ptr<T> removes a reference
to that object, if this was the last Ptr<T> referring to this object,
the latter is de-allocated. Reference counting is implemented as an
Atomic<u64> and is thread-safe.