Remote Call: Use smart pointer for the CallResult returned by Smp::RemoteCall::invokeOn().
This means that the caller to invokeOn() is no longer responsible to
free the returned CallResult. A caller that is not interested in either
waiting for the remote call or getting its result can simply ignore the
Ptr<CallResult> returned by invokeOn() without risking a memory leak.

Implementation-wise, the cpus now use Ptr<CallDesc> in their per-cpu
data.