Testing: Add macro to skip a test when running on a single-core machine.
This new macro TEST_REQUIRES_MULTICORE checks if the machine is a
single-core and if so skips the current test by returning the new
TestResult value: Skip.

Skipped tests do not count as failures. The summary of tests now prints
the number of skipped tests, if any.

Updated existing test cases that required multi-core.