Stack: Re-work how stack allocation interface.
Describes a stack that has been allocated in kernel virtual memory. This
class uses RAII in the sense that the destructor automatically
de-allocate the virtual memory used by this stack.
A Stack instance has ownership of the memory it covers. As such the type
is non-copyable and non-copy-assignable to avoid having multiple Stack
instances referring to the same memory.

Updated code manipulating stacks to use the new interface.