Process: Add first definition of the Proc class.
In this first implementation, processes are running in ring 0 using the
same address space as used to boot the kernel. The Proc class contains
the necessary logic to create a process that runs a kernel function and
perform context switching between processes.

Added new tests for starting a process and performing context switches
between multiple processes.