Testing: Avoid memory leak when restarting an application processor during a test.
The finalizeApplicationProcessorStartup() function would keep a
Ptr<Memory::Stack> of the cpu stack in the call frame before calling
switchToStack(). Since switchToStack() does not return, the destructor
of this Ptr<> was never called and as such the reference counting never
decreased. This meant that if the cpu was ever restarted and a new stack
allocated, the old stack would never be de-allocated.